+++ packages/adapters/ethers5/src/client.ts
@@ -10,6 +10,7 @@ import {
   ParseUtil,
   UserRejectedRequestError
 } from '@reown/appkit-common'
+import { ConstantsUtil, PresetsUtil } from '@reown/appkit-common'
 import {
   AdapterBlueprint,
   AssetController,
@@ -17,6 +18,7 @@ import {
   type CombinedProvider,
   type Connector,
   type ConnectorType,
+  ConnectorUtil,
   CoreHelperUtil,
   OptionsController,
   type Provider,
@@ -27,8 +29,7 @@ import {
   getPreferredAccountType
 } from '@reown/appkit-controllers'
 import { ProviderController } from '@reown/appkit-controllers'
-import { ConnectorUtil } from '@reown/appkit-scaffold-ui/utils'
-import { ConstantsUtil, HelpersUtil, PresetsUtil } from '@reown/appkit-utils'
+import { HelpersUtil } from '@reown/appkit-utils'
 import {
   type Address,
   BaseProvider,
@@ -389,6 +390,9 @@ export class Ethers5Adapter extends AdapterBlueprint {
           this.addConnector({
             id,
             type,
+            explorerId:
+              PresetsUtil.ConnectorExplorerIds[info.rdns || ''] ??
+              PresetsUtil.ConnectorExplorerIds[info.name || ''],
             imageUrl: info?.icon,
             name: info?.name || 'Unknown',
             provider,